move previous mailers to 'mailers/class/action/...
