+++ app/models/user.rb
@@ -65,7 +65,7 @@ class User < ApplicationRecord
   private
 
   def send_welcome_email
-    UserMailer.welcome(self.id).deliver_now
+    Users::Creation::UserMailer.welcome(self.id).deliver_now
   end
 
   def subscribe_to_newsletter
