+++ app/views/mailers/journeys/creation/journey_mailer/confirmation_of_booking.html.erb
