--- app/views/user_mailer/welcome.text.erb
@@ -1,3 +1,0 @@
-
-
-
